Remove some unused imports. Add support in the abstract protocol server and ProtocolServer interface for totalBadRequests and totalErrors
