+++ src/main/java/no/ntnu/okse/protocol/AbstractProtocolServer.java
@@ -33,13 +33,48 @@ import org.apache.log4j.Logger;
  */
 public abstract class AbstractProtocolServer implements ProtocolServer {
 
-    protected String protocolServerType = null;
+    protected String protocolServerType;
+    protected int totalRequests, totalMessages, totalBadRequests, totalErrors;
+
+    public AbstractProtocolServer() {
+        protocolServerType = "";
+        totalMessages = 0;
+        totalRequests = 0;
+    }
 
     // Logger singleton
     protected Logger log;
 
     // Server wrapping thread
     protected Thread _serverThread;
+
+    /**
+     * Total amount of requests from this WSNotificationServer that has passed through this server instance.
+     * @return An integer representing the total amount of request.
+     */
+    public int getTotalRequests() {
+        return totalRequests;
+    }
+
+    /**
+     * Total amount of messages from this WSNotificationServer that has passed through this server instance.
+     * @return: An integer representing the total amount of messages.
+     */
+    public int getTotalMessages() {
+        return totalMessages;
+    }
+
+    /**
+     * This interface method must return the total amount of bad requests recieved by the protocol server.
+     * @return An integer representing the total amount of recieved malformed or bad requests
+     */
+    public int getTotalBadRequests() { return totalBadRequests; };
+
+    /**
+     * This interface method must return the total amount of errors generated by the protocol server.
+     * @return An integer representing the total amount of errors in the protocol server.
+     */
+    public int getTotalErrors() { return totalErrors; };
 
     // Initializer
     protected abstract void init(Integer port);
