+++ src/main/java/no/ntnu/okse/protocol/ProtocolServer.java
@@ -45,6 +45,18 @@ public interface ProtocolServer {
     public int getTotalMessages();
 
     /**
+     * This interface method must return the total amount of bad requests recieved by the protocol server.
+     * @return An integer representing the total amount of recieved malformed or bad requests
+     */
+    public int getTotalBadRequests();
+
+    /**
+     * This interface method must return the total amount of errors generated by the protocol server.
+     * @return An integer representing the total amount of errors in the protocol server.
+     */
+    public int getTotalErrors();
+
+    /**
      * This interface method must implement a complete initialization and startup process of a protocol server.
      * As it is used in the Core Service to fire up all registered protocol servers upon application start.
      */
